requirements.txt
selenium==4.4.3
pytest
python-decouple
allure-pytest